Remove info about verification of the image size
As this guide is no longer relying on the completely empty raw file but
rather on the qcow image with no content inside, the size of the image
in Glance is not 0 bytes anymore and there is no point to check that
really. So this patch removes that point from the guided solution.

Related: #OSPRH-14077